[ONNX] Refactor test_op_consistenct.py and test_fx_op_consistency.py
[ghstack-poisoned]
